Tools: Moved assembler to tools/ folder
