+++ tools/assembler/index.js
