+++ tools/assembler/src/build.js
