+++ tools/assembler/src/defaultOptions.js
