+++ tools/assembler/src/dependencies.js
