+++ tools/assembler/src/process.js
