+++ tools/assembler/src/templates/umd-module.txt
