+++ tools/assembler/src/templates/umd-standalone.txt
