+++ tools/assembler/src/templates/wrap-module.txt
