+++ tools/assembler/src/utilities.js
