+++ tools/assembler/test/build.js
