+++ tools/assembler/test/defaultOptions.js
