+++ tools/assembler/test/dependencies.js
