+++ tools/assembler/test/process.js
