+++ tools/assembler/test/test-main.js
