+++ tools/assembler/test/utilities.js
