+++ tools/compareFilesize.js
@@ -3,7 +3,7 @@ const { getBuildScripts } = require('./build.js');
 const { compile } = require('./compile.js');
 const { sync: gzipSize } = require('gzip-size');
 const { getFile, writeFilePromise: writeFile } =
-    require('../assembler/src/utilities.js');
+    require('./assembler/src/utilities.js');
 const { join, resolve } = require('path');
 
 // TODO: write JSDoc
