+++ tools/compile.js
@@ -6,7 +6,7 @@ const statSync = require('fs').statSync;
 const {
     getFile,
     writeFile
-} = require('../assembler/src/utilities.js');
+} = require('./assembler/src/utilities.js');
 const colors = require('colors');
 
 const compileSingleFile = (path, sourceFolder, createSourceMap) => {
